Move `ScalarValue` tests alongside implementation, move `from_slice` to `core`
